Experiment with miri on the linux_raw backend.
Add a hack to use libc calls instead of inline asm at the very bottom of the
linux_raw backend, so that we can test most of the linux_raw code under miri.

This is based on the strict-provenance-lints branch.

To test this, run:

`env MIRIFLAGS="-Zmiri-strict-provenance" cargo +nightly miri run --example hello --features=stdio`